% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget_execution}
\alias{budget_execution}
\title{Budget Execution Data}
\format{
A data frame with the following columns:
}
\source{
BOOST World Bank Data \url{https://www.worldbank.org/en/programs/boost-portal/boost-data-lab#1}.
}
\usage{
budget_execution
}
\description{
A dataset containing budget execution information, including approved and executed budget values, execution ratios, and classifications by region, sector, and spending type.
}
\keyword{datasets}
