% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigoanalytics.R
\name{compute_cluster_diff}
\alias{compute_cluster_diff}
\title{A function to Compute Change in Cluster Values Between Two Years for Each Country}
\usage{
compute_cluster_diff(data, cluster_name, from_year, to_year)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{cluster_name}{The name of the cluster to filter.}

\item{from_year}{The starting year for comparison.}

\item{to_year}{The ending year for comparison.}
}
\value{
A data frame with the computed differences.
}
\description{
Computes the difference in values for a specific cluster between two years.
}
