% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigoanalytics.R
\name{compute_regional_statistics}
\alias{compute_regional_statistics}
\title{A function to Compute Regional Min/Average/Max Statistics for a Given Cluster}
\usage{
compute_regional_statistics(data, dimension)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{dimension}{The column name (as a string or symbol) representing the dimension to compute statistics for.}
}
\value{
A data frame with computed regional statistics.
}
\description{
A function to Compute Regional Min/Average/Max Statistics for a Given Cluster
}
