% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dgss}
\alias{dgss}
\title{GovTech Maturity Index}
\format{
A data frame with 198 rows and 21 variables:
\describe{
\item{\code{code}}{character Country code}
\item{\code{economy}}{character Economy name}
\item{\code{level}}{character Income group level}
\item{\code{population}}{double Population}
\item{\code{gni}}{double Gross National Income}
\item{\code{gnipc}}{double Gross National Income per capita}
\item{\code{region}}{character Region}
\item{\code{tmis_sta}}{double Tax MIS Status}
\item{\code{cust_sta}}{double Customs MIS Status}
\item{\code{hrm_sta}}{double HRMIS Status}
\item{\code{payr_sta}}{double Payroll MIS Status}
\item{\code{e_prc_sta}}{double e-Procurement Status}
\item{\code{dms_sta}}{double Debt Management System Status}
\item{\code{pims_sta}}{double Public Investment MIS Status}
\item{\code{tmis_yr}}{double Tax MIS year established}
\item{\code{cust_yr}}{double Customs MIS year established}
\item{\code{hrm_yr}}{double HRMIS year established}
\item{\code{payr_yr}}{double Payroll MIS year established}
\item{\code{e_prc_yr}}{double e-Procurement year established}
\item{\code{dms_yr}}{double Debt Management System year established}
\item{\code{pims_yr}}{double Public Investment MIS year established}
}
}
\source{
https://datacatalogfiles.worldbank.org/ddh-published/0037889/DR0065450/WBG_GovTech\%20Dataset_Dec2020.xlsx
}
\usage{
dgss
}
\description{
The 2020 GovTech dataset contained data/evidence collected from government websites using remotely measurable indicators (due to the COVID-19 pandemic) mostly reflecting de jure practices.
}
\keyword{datasets}
