% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dynamic_ctf_clusters}
\alias{dynamic_ctf_clusters}
\title{Dynamic CTF Clusters Data}
\format{
A data frame with the following columns:
\describe{
\item{cluster}{Description of the cluster variable.}
\item{country_code}{Country Code ISO3.}
\item{country_name}{Country Name.}
\item{income_group}{Income WB Classification.}
\item{lending_category}{Lending WB Classification.}
\item{region}{WB Region.}
\item{value}{Indicator Coefficient.}
\item{year}{Year.}
}
}
\source{
CLIAR World Bank Data.
}
\usage{
dynamic_ctf_clusters
}
\description{
A dataset containing information about dynamic CTF clusters.
}
\keyword{datasets}
