% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigoanalytics.R
\name{generate_regional_minmax_plot}
\alias{generate_regional_minmax_plot}
\title{A function to Generate a Min/Average/Max Labeled Plot by Region}
\usage{
generate_regional_minmax_plot(data, dimension)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{dimension}{The column name (as a string or symbol) representing the dimension to filter and plot.}
}
\value{
A ggplot object with the generated labels plot.
}
\description{
Creates a plot with labels for a specific dimension.
}
