% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{static_ctf_clusters}
\alias{static_ctf_clusters}
\title{Static CTF Clusters Data}
\format{
A data frame with the following columns:
\describe{
\item{country_code}{Country Code ISO3.}
\item{country_name}{Country Name.}
\item{income_group}{Income WB Classification.}
\item{lending_category}{Lending WB Classification.}
\item{region}{Region WB Classification.}
\item{value}{Indicator Coefficient.}
\item{var_name}{Indicator Name.}
\item{variable}{Indicator Code.}
}
}
\source{
CLIAR World Bank Data.
}
\usage{
static_ctf_clusters
}
\description{
A dataset containing information about static CTF clusters.
}
\keyword{datasets}
