% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigoanalytics.R
\name{wrap_facet_titles}
\alias{wrap_facet_titles}
\title{A function to wrap graph's text labels}
\usage{
wrap_facet_titles(x, width)
}
\arguments{
\item{x}{object or variable that need to be wrapped}

\item{width}{number of characters}
}
\description{
A function to wrap graph's text labels
}
